Export more methods that will be needed to develop the global installs
I made a copy of the VersionResolver, eventually all of that code will be deleted.
So I would not worry about it.
The initial thought was it'd be better to have a 2nd version resolver that handled more types of inputs.
Eventually I just expanded the existing one and dumped this file.