Make all of the linux version resolver tests pass
We needed to add the mock executor to the mock distro provider so we could see if it tried to run the install command or the update command. I also added a test to see if things are rejected (when you try to install a lower major.minor than what's already on the machine) but we should see if we want that behavior by checking in with the PMs on the team!