Implement tests for Global Installer Resolver. Rename classes to be sdk and runtime agnostic.
Since classes will be running a global install of the sdk and the runtime in the future, it makes sense to rename them now. The word SDK also makes them more verbose.

This implements the tests for the global installer resolver. It also makes improvements to said resolver I observed while testing. We never needed, to check for mac conflicting installs as they would be caught by the mac installer.

I added a file for the windows and mac installer tests that will be implemented soon.