Fix the PATH Logic for Global Installs and E2E Test.
The e2e test has added an option to not actually do the install via the environment variable `VSCODE_DOTNET_GLOBAL_INSTALL_FAKE_PATH`. We dont want to run the install as it'd put things on our build machines and quickly escalate into a disaster for other repositories.

We now set the PATH for a global install because there is no linux installer that sets the path.

This also includes minor improvements to other tests like using assert.includes instead of just asserting is true and checking substring.includes.