Fix Installer Status Code Handling + Partial Installs
We do not have logic to do an uninstall for the global install yet, but it's good to be able to check for a partial install and log it nontheless, which is what we do here.

We fixed the windows installer to use the installer status exit code instead of stdout because the windows installer does not print to stdout or stderr. We also dont want to fail on reboot success codes.

We also cleaned up code by removing unneeded includes. Moving a test string into a file.