Fix 32-bit bugs
I thought x86 was ia32 but that is not correct.
the registry key lookup was not working on 32 bit machines
node could not resolve system root
the install location was different than I expected and the registry keys are stored in a different location
in addition, during refactor for the linter I forgot to remove the ' in one of the strings causing failure
Fix Installation Validation to not expect an .exe because linux doesnt use .exe or any 1 dotnet file
todo: see on win32 why an existing install does not return 0. It might be because of the reg key error