Fix bugs related to version and error parsing
EventStreamEvents need to be created with an error; you cannot throw the
... event stream event. Instead we need to call .error to throw their error.
This is not a great pattern. I chose it to avoid breaking existing code patterns.

This also fixes an issue with the Debugging library causing the runtime extension
to be bundled into the sample extension. this would cause the sample extension to fail.