Add copy step for library dependent files
The library does not get webpacked nor should it be as its not an extension
But it needs the files to be copied for it to access them correctly as it would be in the .VSIX file
We should be absolutely sure and verify with manual test that the VSIX when packed has the same behavior as it will load differently.
But this is the best solution I found; install scripts folder is never accessed in a lib test to my knowledge, so it never encountered this issue.