Verify that PowerShell Install Scripts can Execute
Add an error to tell users how to install powershell if it's not on the PATH.
Also check execution policy to see if the elevation level needed to run dotnet install is available or not.

Also changes the timeout value for those who have slower internet -> assuming 8 mbps/sec for downloading .NET is roughly 4.5 minutes or 270 seconds + internet SP ramp up time.

Changes messaging for timeout to include instructions on how to change timeout value and also makes button more likely to be clicked on, or that's the hope.