Fix Mac OSX Related Bugs
The / in cp is not allowed on mac even though it works in linux so it was removed

We used the x64 path but that seems to be for arm only machines. Confirming this following the blame of who wrote the docs on dotnet install paths : https://github.com/dotnet/docs/pull/27186

Increased timeout value on tests.
Ideally, after downloading the .NET SDK if an existing install is detected we should print that we are no longer going to do an install.