Remove Legacy Only After Completion
Fixes SDK tests
Improves event variable names to match what they are actually getting

A few events still use version. This may be a bit annoying with the telemetry, because the beginning events are before architecture is determined, so they use a different version key vs the install key. I worry this may make it harder to drive failure insights. We could change this. but I left it because this is technically an existing problem -- the version used in these events is the unresolved version, so it will not be the fully specified version used in later events.

I moved the removal of legacy installs down incase someone is offline, then we keep their legacy install if the new install failed to reduce breakage chance