Use a CommandType instead of strings
Per @baronfel suggestions, we should avoid parsing the commands ( I agree ) as it will cause a lot of pain and potential bugs. This updates the api surface and distro json file to use an array of strings instead of one string. Its uglier to call into but more risk-adverse (well, at least once this is as well tested as the old code)