Fix Axios Storage Handler to Handle Timeout & TLS Handshake Properly
https://axios-cache-interceptor.js.org/guide/storages The storage handler is supposed to prevent timed out requests from going into the cache per the documentation which it does not do.

Also, VS Code will try to unwrap a circular reference which the request will be circular in the TLS handshake. This is a 'bug' in serializeRequestArguments in extensionHostProcess.js from VS Code. This will prevent us from removing from the extension state in the future as trying to access it will cause an exception and cancel the request