Switch to memory based storage
This means the cache will not persist over different vscode windows, which is not optimal. However, until vs code does not fail with circular json references or non json objects with this cache, we do not seem to have ways to filter what goes into the cache and what does not. Even for one request, the no cache method still ends up putting things into the storage, see: https://axios-cache-interceptor.js.org/config/request-specifics#cache. Then, the request object given to set does allow us to filter some of the content per request, but even then, one request seems to cause multiple potential cache entries. In addition, I looked at the code of the axios-cache-interceptor- and the loading state never gets updated until something is in the cache, so my previous code was based on incorrect assumptions, even if the logic would have worked if we had that kind of graunlar control