Improve timeout error handling
Axios already has its own timeout, so we dont need a 2nd timeout handler,  I was able to confirm it works by using https://httpstat.us/524, and it does report the time it took to timeout in a similar fashion. We will get more information from errors this way when it is available.