Account for versions of linux that we dont have mapped
The theory is a version with the same major but with the maximum lower minor will be the closest in behavior, then the maximum version that is lower than the current version is most likely to have the same behavior. If there is no smaller known version, it will just use the minimum version that is higher.

This logic may not always work. But with the given circumstances, we are covered by including the first version where preinstall commands were not needed