Fix unit tests expecting old data
The code changes that adopted the event stream were unnecessary and complicated. It is not that hard to update the version and arch of the same worker and re use it, in fact not doing so defeats the purpose
of these tests in someways.

More importantly we safe guard against an undefined proxy setting in the web worker.