Fix existing tests that worked for a long time by pure magic
- The acquisition promises were tracking 2 things...
1 is the current working in progress installs.
2 is all installs that were ever done, with the assumption no outside source will ever delete the installs, which is incorrect.

When fixing 2, several tests broke. This is because they relied on the promises to be there even though nothing was installed. We dont actually want to install on the test so we dont change that, instead we add a boolean to fake the thing into thinking an install was laid onto the disk