Protect against no versions of .NET to recommend  (#1768)
* Update Dependencies

* Remove is online

Is-online now requires ES6.
It has not been maintained in over 2 years.
The checks it does are fairly rudimentary. We only rely on this code to try to diagnose the error which has already happened.
I would trust this code rather than rely on external libraries.

* Fix timeout logic for no is online

* Fix Linting Issues

* Protect against no versions of .NET to recommend

The API contract is to return an IDotnetVersion which is an array of versions. If there are no packages of .NET available to install, we should have a specific message for that and return an empty array instead of what we do right now, which is

a - returning a bad version
b - returning undefined
c - erroring out

* respond to linter to accept any error type