Add global shared file path for extension (#1726)
* add logic for adding a global shared path to extension

* add tests for global path config

* update comment out code

* update mock comment

* remove commented out code in ExtensionConfigurationWorker

* update ExistingPathResolver

* add tests for global shared path

* remove empty line

* resolve duplicate import

* fix tslint

* resolve format comments

* get rid of set global path to avoid type error

* add logic for checking local paths first in ExistingPathResolver

* create set global path method

* fix test

* Fix Typo

* change the field name for global shared path; add sharedExistingDotnetPath to package.json; change get and set method for shared path for ExtensionConfigurationWorker

* fix typo

* update readme for global shared path

* Add more examples to the path for .NET

* Consolidate read.me section to one page

* Simplify Logic for Failure to set Manual Path

I thought about the direction this should go, and I think that the old logic only makes sense because we'd need to concat the existing settings. In this case we are setting the global setting, so we can just add it without having to do any additional logic.

* Consolidate setting global and local to one function for simplicity

* update sample page to use the string instead of path obj

* Refactor path setting functions to two separate ones

We need to access the configuration with the direct objects the users can use the settings json file, hence this could not be simplified to a shared object.

* update class member names to disambiguate overloaded terms

* Check whether a matching extension id extension exists before calling the global path

* Make read.me more clear

* Update read.me to show that the example is only an example

* .some may not exist in certain versions so dont use it?

* Dont activate the runtime extension under test

* dont rely on test env variable since we cant set one within vscode launch profile

* Fix ability to run extension tests under launch json

We activate the extension in the test to inject custom objects to track 'telemetry' under test and analyze the telemetry events to ensure things are going well.

We cant activate the extension twice.
In the recent past we made a change to auto activate as a necessity to check the operating system before exposing APIs to users that were not available on their OS.

When I made that change I think we actually broke the ability to run the runtime extension tests under the launch profile. This fixes that.

* Remove global setting from test

This setting impacts the return path of all tests when correctly implemented. We dont want to dupe the extension test configuration for a single test to run it with a different configuration either, so we cant test it this way.

* Add tests that dont rely on the extension

* fix typo

* improve warning messages to correctly diagnose what went wrong

* remark that you need to restart vscode for the setting to take effect

* return empty properties array instead of undefined so properties are forwarded via telemetry observer

---------

Co-authored-by: Noah Gilson <noahgilson@microsoft.com>
Co-authored-by: Noah Gilson <nagilson@umich.edu>