Track which Extensions Own Which Installs (#1748)
* WIP - Migrate Install Key strings to type to allow future flexibility with installs -- e.g. non-versioned install folder names so we can auto update. Start collecting more information during install about who installed it so we can ref count them for uninstalls. Does not migrate telemetry events or other code yet which needs to be experimented on further

* WIP - Migrate all code to use InstallKey Type & Fix Bugs

There are several big things here that might seem like small code changes.

1 - Before when we found a 'partial install' we would delete ALL local runtimes. I am astounded this exists. There are other extensions that might depend on those other runtimes and this just deletes it from underneath them. A partial install is rare, I think you need to pkill vscode or something since the install script runs out of proc. We consider an install partial if our code never got to mark it as installed after calling the install script. Its rare but does happen, at 2010 occurences out of 470k in recent days.

Renames a function which appears to handle partial installs to clarify it only manages local ones.

Gets rid of some unused functions or moves them to a better place.

Renames IInstallKey to DotnetInstall. The install key is the string representing the directory of the install. It shouldn't have two meanings.

Add logic for ref counting which extensoins own which installs.

* wip - record extension owner for installs

Fix record keeping to properly ref count extensions
Rename types to match standard better

Still need to figure out how to manage uninstalling and 'graveyard' as well as how to remove a 'legacy' install via version when it is a dotnetinstall object.

* Separate out 'graveyard' code or management of old installs to another class

* Dont pull files from dist - bad copilot, bad

* Update some more of the code to use new types

* update mock objects to get the build working

* respond to linter

* Add back in the old telemetry prooperties

I got rid of them because they are duplicated data. But TBH there are a lot of things that rely on the pre-existing property names. It would be much harder to query new and old data together if getting rid of them, so I added them back even though it causes some data duplication.

* Fix bugs with install management

* add missing semicolon lol

* Fix a bug with clearing out the installed list

The logic was missing a !.

* Fix Install Graveyard Logic

* Fix test logic for legacy installs

* Skip unnecessary preinstall local sdk check on global sdk

* Add back code to account for preinstall in sdk status check

* Remove unnecessary string checks

We never give a string to this class anymore so it's not needed

* Add event for when tracking installs messes up somehow

* Refactor classes out to many more files

* Respond to linter

* Isolate install tracking management to a separate class

* Clean up acquisition worker and remove unnecessary logic

* Port the worker to using an external install manager to isolate protected data

* Respond to linter concerns

* Remove potential deadlock with installation management

* Increase timeouts

With the change where you need to hold a lock to edit extension state to protect against timing challenges where an install is being added or deleted at the same time and can cause incorrect behavior, we had to slow down the product.

* Migrate to a singleton for install tracker so no dupliacte lock is created

* Working build for changing 'isRuntime' to mode for 'aspnet' 'sdk' or 'runtime'

* undo the singleton in test as the context needs to be per test

* remove the singleton pattern as it will result in a bug with multiple acquisition requests having a stale context

* Add Tests for Install Tracking with who owns what install

* improve timeout on test

* Revert "improve timeout on test"

This reverts commit a4168f6ec57c7360cdd3b8e20b35ad792b6ccf6c.

* fix timeout

* Fix Test

* Fix Logic for adding pre existing installs

* Fix tests to be successful

* Code Cleanup & Remove Potential Tech Debt for Modes

remove places where we make assumptions on the existing model (sdk or runtime) to prevent future tech debt

* respond to linter

* Fix bug with context being in a different mode than a legacy installation

Mark old function as deprecated and add a comment explaining why our assumption there works

Fix possible bugs with context management where context could be a different mode than what was previously installed which could cause future bugs down the line that would be hard to track down

* Add back in removed events to prevent breaking old telemetry dashboards

* Final Code Review from Me

* Fix Test

* Remove circular dependency from test artifact

* add missing file header from refactor

* fix test

* delete directory per mode file as there is a git bug where changes in file caps is not being picked up

* Add back into directory file with proper capitilzation per git bug

* Add missing await which might cause test to fail inconsistently

* Make the lock based timeout based on the user setting for timeouts