Update Linux Apt Commands to Handle Lock Timeouts
We also pass the -y flag when it can help push through updates.
Apt will return 100 whenever there is any failure. https://github.com/Debian/apt/blob/aa56836331870d975c212a5df2f13db9ce3914bf/apt-private/private-cmndline.cc#L602

This means the errors we are seeing are not just due to locking errors. But, perhaps this will help.

Also this will try to run update twice, to see if that 2nd try will work. It seems a bit silly, but it does actually seem to improve the chance of success when something else is running on the system.