Version 2.1.1, Hot fix for installation record tracking (#1888)
* Hot fix for installation record tracking

We renamed it from 'key' to 'id' because the string 'key' is filtered out of all log events via vscode because 'key' looks like a secret. We did this in a rush and vendor testing did not have pre-existing machine state to realize the old data structure would still be in the local memory for pre-existing users. This converts the data in memory to the new type.

* Prevent adding corrupted installation records into the data