2.1.4 branding (#1928)
* Fix an issue parsing promises

We changed the function `getPromise` to not use a lock in version 2.1.2. This was a mistake. I thought it would return the unresolved `Promise` object if it was a sync function. That is not the behavior, instead it will skip that line of code and return null even if the promise exists. This causes multiple processes to try to grab the installer file handle.

When one fails to grab the file handle, it causes a chain-reaction of cascading failures for concurrent requests, because that is supposed to be blocked from happening.

In addition, the check I added to kill the sudo process once it is finished runs on all platforms despite the master sudo process only being present on Linux, which has resulted in some higher timeout rates. We need to not do this on other platforms.

* Fix tests

we need to clear the promise out after each test is done

* dont call uninstall all in sdk test, its the wrong command

* reset event stream in the test

* 2.1.4 branding - new version due to vscode marketplace outage causing failure to upload version 2.1.3 correctly