Add new API `dotnet.findPath` to find the .NET host to use if one already exists that meets a requirement. (#1954)
* update PATH setting description to make it more clear

* add very prototype code

* add tests and get runtime check working

* Fix some bugs

* Consider that the .NET SDK can also satisfy the Runtime

* fix typo

* fix the tests

* fix

* fix another test

* Fix tests even more

* undo yarn changes

* fix the final test hopefully

* undo yarn changes

* Dont use dotnet.exe as its not platform agnostic

* use "markdownDescription" for nicer rendering in VSCode

* Tweak call to action to use VSCode mechanisms first, then our standard installation docs, and only in the last resort mention PATH munging.

* Fix URLs

* Move to version 2.1.7

* add basic command'

* Ignore existingPath setting for SDK installs.

C# DevKit never uses the path returned by our installation. This means users would think this path would change the sdk that this extension uses but that is not the case. This path to dotnet.exe is meant to be the path for the runtime for extensions to run on, and not the SDK path. It's confusing that the setting was used for both and a misstep in a way. DevKit is the main caller of this API so we think we can change this with minimal breakage.

* Remove warning setting and fix invalid path setting

The setting must be accessed earlier. This means vscode will need to be restarted. We also update the readme and messaging a bit so its more publicly clear in all places what the setting is for.

* Fix test

* Refactor code out into a Validator for Conditions

dotnet --list-runtimes and more need to be called in more places. This is a separate task so it should be done. I did not change the code in any way except for adding the requirement clause type.

* Prepare code to validate the path

* add a lot of prototypey code

* add comment for future work

* merge with main

* look up the architecture

* Improve the code

* Go 2 Directories Up to find the True Path on PATH

* Final initial loop of API code

* Fix bug parsing list runtimes

* Add tests

* Fix test and search for where if its not installed

* Consider where may return multiple values

* Fix test

* tests mostly working

* code cleanup - get rid of extra api to set env

* Restore the env var so we dont edit it for other processes

* Uncomment the remaining tests

* Respond to lint

* fix callback

* Fix path to be os-gnostic

* Only search for where on windows and also search for which

* provide env to the find command so /usr/bin/whcih can be used

* Call which which instead of which so the correct command can be found

* Install 3.1 instead of 7.0 because the DTL CI machines seem to have a 7.0 SDK on them :zany:

* give up on arch check for now because it is inaccurate, see comment

* Add github issue in comment for context

* make linter happy

* Respond to PR feedback

* Migrate to connection strings

Resolves https://github.com/dotnet/vscode-dotnet-runtime/issues/1958

The old application insights key was created by @LakshanF, when we migrated to the new vscode-extension-telemetry service, their API had a breaking change to require a connection string instead of an insights key. https://github.com/dotnet/vscode-dotnet-runtime/pull/1948

The connection key can be public and is hard coded. Our existing key has been in our open source, source code for many years. Here is what their guidance says: https://www.npmjs.com/package/@vscode/extension-telemetry

> Follow guide to set up Application Insights in Azure and get your connection string. Don't worry about hardcoding it, it is not sensitive.

* Respond to PR Feedback

---------

Co-authored-by: Chet Husk <chusk3@gmail.com>