Condition the name of the build based on OS
Why is windows the bomb emoji and not the windows emoji?
And why did I do this:

The windows build goes last because its in alphabetical order based on the behavior of how devops works. This is bad because it should go first and makes the other steps take longer as they depend on only the windows task. The bomb emoji is one of the first emojis thats not an emoticon before the penguin emoji so I picked that one..