Exclude Microbuild from scanner
Microbuild is failing credscan following TSA upload. It has several issues and the pipeline is publishing artifacts to the root artifacts staging directory and CredScan is considering everything under that path as part of the pipeline output. MicroBuild is being installed there, so it is incorrectly being brought into the scan.

This repo does not own microbuild, we only do it for signing.