Leverage Managed Identity Publish (#1942)
* add publish task

* move template out of steps and into jobs

* get rid of pool

* dont pass 1es pool

* add depends on chain

* use vsce latest

* run package on publi

* add approval step:

* vsixs are stored in the root, so don't check in the extension folder

* Switch to MI instead of a PAT

* Fix ps1 issue

* Only deploy on internal

* Use pool vs vmimage

* Try to fix pool condition

* Use a different pool

* Condition the name of the build based on OS

Why is windows the bomb emoji and not the windows emoji?
And why did I do this:

The windows build goes last because its in alphabetical order based on the behavior of how devops works. This is bad because it should go first and makes the other steps take longer as they depend on only the windows task. The bomb emoji is one of the first emojis thats not an emoticon before the penguin emoji so I picked that one..

* It's pretty silly you cant define a variable in a template but here we are

* Maybe it has to be in plaintext

* The deployment step is not necessary

* wait for wait job

* Set the emoji

* Move windows to the top

* Set 'use 1es pool' to true

* Switch to a different ordering so windows runs first, then linux

* try to fix bug where version is not passed and publish args is wrong

* Try to pass var into another job since we cant have a template in the jobs

* Revert "Try to pass var into another job since we cant have a template in the jobs"

This reverts commit 7bdf2dfec95f13f053bc016cb767be9257da744c.

* Try to condition the yamls omewhere else

* is yaml failing

* Revert "is yaml failing"

This reverts commit b014dd042fcc63972477a087033ae304c8639a14.

* Revert "Try to condition the yamls omewhere else"

This reverts commit 7ccb36f01c00634eed03098b65eaa0d4a5d72427.

* Revert "try to fix bug where version is not passed and publish args is wrong"

This reverts commit 592349be71605208003ccebe23f271cdde8d1cef.

* Get version again

We cant return a value from a template
We cant make a template callable with jobs in another job
We cant pass a variable from one job to another if its a job in a job
We cant do so many things in yaml

* yaml indentation

* Don't skip checking out the repo, we kinda need that to get the version