Show Cancellation Style Errors (#1989)
I made a change a while back such that 'cancelled request' style errors dont cause a pop up because I figured they were not useful. The vendors reported it as a bug. Do you think we should keep a pop up for those types of errors?

The types of errors include: entered a bad version, you cancelled the password prompt or admin prompt, you have a custom linux install or conflicting feeds, your distro is not supported, etc

Resolves https://github.com/dotnet/vscode-dotnet-runtime/issues/1826

If this is closed, close this issue.
Waiting for PM feedback on @baronfel. I think we can display it.

Extensions can set errorConfiguration in the acquire context

export declare enum AcquireErrorConfiguration {
    DisplayAllErrorPopups = 0,
    DisableErrorPopups = 1
}
to ignore this.