Consider when the user Cancels the SDK Installer on Mac (#1993)
* Try asking the user if the want to continue

* Add a message on mac if someone tries to cancel the installation

Consider that to tell if the user wanted to say 'no' or 'yes' to retry, that doesnt indicate whether they meant to do the install or not. We would then need more options to tell if we actually failed or not. I dont want to rely on this and I dont think it is a high enough priority with the complexity required to reroute the code to add a retry button that does a certain retry depending on if someone clicked to cancel the install on mac

* Consider if the dotnet file doesnt exist but the folder does

* Fix bug