Fix .NET Version Parsing for FindPath API (#2000)
* Fix Version Parsing for FindPath API

The old comparison used string comparison which is fine for versions of dotnet until dotnet 10. '10.0' < '9.0'. :joy: This fixes that. I suppose it's a bit clunky and we could dedupe / share the compare logic but it's not being used elsewhere in this same form, so I don't plan to do that right now. There are also places that depend on the output versions being strings, so that's a problem to change that now too. It probably should have been a class from the beginning but unfortunately that just isn't how it was initially done, which is understandable because the scope at that time was much more limited.

* Fix some of the code

* Fix test

* Dont comment out the other tests

* update comment