Consider Runtime Patch Versions in PATH API
The C# extension will now fail if the runtime is not 8.0.10 or higher on mac so we need to support patch version lookup.

Sadly we cannot migrate to semver because semver rejects strings such as 8.0 which are allowable in many parts of our code and APIs.

This also adds code for SDK lookup but no test since that's harder to test, like remarked in the code.