Fix acquisition error on non fully qualified version
The new, unreleased code for getting a feature band of a sdk version try catches when it tries to get the band. It will not throw if there is no band in the string which can be expected in certain cases, but it still posts an error to the event stream. Posting an error to the event stream causes the error to bubble up to the user even if it is caught. We need to take some additional action here so the user doesn't see an error popup for something that's not an error. This is blocking our release with all of the findPath features.

Resolves https://github.com/dotnet/vscode-dotnet-runtime/issues/2015

Version updates are to update the shas of the internal packages which got updated.